Changes to type checking in transformations, renaming of composition layer

+++ spatial_transformations/config.py
@@ -8,4 +8,4 @@
 
 
 import numpy
-DTYPE = numpy.float64
+DTYPE = numpy.float32
